refactor(image): introduce image base and image layer; rename vrange to v_range
